pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "chromadb>=1.3.7",
    "google-generativeai>=0.8.6",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.52.2",
    "supabase>=2.0.0",
    "watchdog>=6.0.0",
]
